[fix] template opc 적용
