+++ src/domain/template/service/handler/TemplateNotifier.ts
@@ -1,0 +1,154 @@
+import { Injectable } from '@nestjs/common';
+import { TemplateHandler } from './TemplateHandler';
+import { TemplateOperation } from '../types/Operation';
+import { TemplateContent } from '../../dto/response/TemplateContent';
+import { UserTemplate } from '../../domain/entity/UserTemplate';
+import { Affiliation } from 'src/domain/user/domain/entity/Affiliation';
+import { sortCompanyPublic, sortCompanyPublicArray } from '../../util/data';
+import { formatDate } from '../../util/date';
+import { Question } from 'src/domain/challenge/domain/entity/Question';
+import { UserChallenge } from 'src/domain/user/domain/entity/UserChallenge';
+import { UserTemplateHelper } from '../../helper/UserTemplate.Helper';
+import { UserApi } from '../../infrastructure/User.Api';
+import { DataMapperService } from '../../domain/service/DataMappper.Service';
+import { ChallengeApi } from '../../infrastructure/Challenge.Api';
+import { TemplateInformation } from '../../dto/response/TemplateInformation';
+import { UserVerifyService } from 'src/global/exception/user/UserVerify.Service';
+
+@Injectable()
+export class TemplateNotifier
+  implements
+    TemplateHandler<[number, string, number], Promise<(GetCommentNotify | GetLikeNotify)[]>>
+{
+  operation: TemplateOperation = 'SEND_NOTIFY';
+
+  constructor(
+    private readonly userApi: UserApi,
+    private readonly userTemplateHelper: UserTemplateHelper,
+  ) {}
+
+  async handle(request: [number, string, number]): Promise<(GetCommentNotify | GetLikeNotify)[]> {
+    const [userId, organization, challengeId] = request;
+    const userChallengeAndAffiliationData =
+      await this.userApi.requestUserChallengeAndAffiliationByChallengeIdWithUserIdAndOrganization(
+        challengeId,
+        userId,
+        organization,
+      );
+    const userTemplateAndCommentAndLikeData =
+      await this.userTemplateHelper.giveUserTemplateAndCommentAndLikeByUserChallengeId(
+        userChallengeAndAffiliationData.getId(),
+      );
+    const hasNoData =
+      !userTemplateAndCommentAndLikeData ||
+      userTemplateAndCommentAndLikeData.length === 0 ||
+      !userTemplateAndCommentAndLikeData.some(
+        (item) => item.comments && item.comments.length > 0,
+      ) ||
+      !userTemplateAndCommentAndLikeData.some((item) => item.likes && item.likes.length > 0);
+    return hasNoData
+      ? []
+      : this.proccessNotifyData(userTemplateAndCommentAndLikeData, userChallengeAndAffiliationData);
+  }
+
+  private async proccessNotifyData(
+    userTemplateAndCommentAndLikeData: UserTemplate[],
+    userChallengeAndAffiliationData: UserChallenge,
+  ) {
+    const extractAffiliationId = this.extractAffiliationIdAccordingToCommentAndLike(
+      userTemplateAndCommentAndLikeData,
+    );
+    const [commentAffiliationData, likeAffiliationData] = await Promise.all([
+      this.userApi.requestAffiliationById(extractAffiliationId.commentAffiliationIds),
+      this.userApi.requestAffiliationById(extractAffiliationId.likeAffiliationIds),
+    ]);
+    const sortedComment = this.makeCommentShapeAccordingToUserTemplate(
+      userTemplateAndCommentAndLikeData,
+      userChallengeAndAffiliationData,
+      commentAffiliationData,
+    );
+    const sortedLike = this.makeLikeShapeAccordingToUserTemplate(
+      userTemplateAndCommentAndLikeData,
+      userChallengeAndAffiliationData,
+      likeAffiliationData,
+    );
+    const mergedCommentAndLike = this.mergeAndSortTimeCommentAndLike(sortedComment, sortedLike);
+    return mergedCommentAndLike;
+  }
+  private extractAffiliationIdAccordingToCommentAndLike(userTemplate: UserTemplate[]) {
+    const commentAffiliationIds: number[] = userTemplate.flatMap((userTemplate) =>
+      userTemplate.comments.map((comment) => comment.getAffiliationId()),
+    );
+    const likeAffiliationIds: number[] = userTemplate.flatMap((userTemplate) =>
+      userTemplate.likes.map((like) => like.getAffiliationId()),
+    );
+    return { commentAffiliationIds, likeAffiliationIds };
+  }
+  /**
+   *
+   * @param userTemplate 유저 챌린지에서 작성한 템플릿 데이터
+   * @param userChallengeAndAffiliationData 유저 챌린지와 소속 데이터가 포함된 데이터
+   * @returns 유저가 자신의 템플릿에 단 댓글이 제거된 데이터
+   */
+  private makeCommentShapeAccordingToUserTemplate(
+    userTemplate: UserTemplate[],
+    userChallengeAndAffiliationData: UserChallenge,
+    affiliation: Affiliation[],
+  ): GetCommentNotify[] {
+    return userTemplate.flatMap((userTemplate) =>
+      userTemplate.comments
+        .filter(
+          (comment) =>
+            comment.getAffiliationId() !== userChallengeAndAffiliationData.affiliation.getId(),
+        )
+        .map((comment) => {
+          const matchedAffiliation = affiliation.find(
+            (affiliation) => affiliation.getId() === comment.getAffiliationId(),
+          );
+          return {
+            commentId: comment.getId(),
+            content: comment.getContent(),
+            createdAt: comment.getCreatedAt(),
+            sign: comment.getCheck(),
+            userTemplateId: userTemplate.getId(),
+            templateName: userTemplate.getTemplateDate(),
+            nickname: matchedAffiliation.getNickname(),
+            type: 'comment',
+          };
+        }),
+    );
+  }
+  private makeLikeShapeAccordingToUserTemplate(
+    userTemplate: UserTemplate[],
+    userChallengeAndAffiliationData: UserChallenge,
+    affiliation: Affiliation[],
+  ): GetLikeNotify[] {
+    return userTemplate.flatMap((userTemplate) =>
+      userTemplate.likes
+        .filter(
+          (like) => like.getAffiliationId() !== userChallengeAndAffiliationData.affiliation.getId(),
+        )
+        .map((like) => {
+          const matchedAffiliation = affiliation.find(
+            (affiliation) => affiliation.getId() === like.getAffiliationId(),
+          );
+          return {
+            likeId: like.getId(),
+            createdAt: like.getCreatedAt(),
+            sign: like.getCheck(),
+            userTemplateId: userTemplate.getId(),
+            templateName: userTemplate.getTemplateDate(),
+            nickname: matchedAffiliation.getNickname(),
+            type: 'like',
+          };
+        }),
+    );
+  }
+  private mergeAndSortTimeCommentAndLike(comments: GetCommentNotify[], likes: GetLikeNotify[]) {
+    const mergedArray: (GetCommentNotify | GetLikeNotify)[] = [...comments, ...likes];
+    mergedArray.sort((a, b) => {
+      return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
+    });
+    return mergedArray;
+  }
+}
